removes some try/except blocks from controllers

+++ controllers/users_controller.py
@@ -5,7 +5,7 @@ from uuid import UUID
 from db import db
 from models.app_users import AppUsers, app_user_schema, app_users_schema
 from util.reflection import populate_object
-from lib.authenticate import auth, validate_uuid4
+from lib.authenticate import auth, validate_uuid4, authenticate_return_auth
 
 
 def user_add(req):
@@ -20,17 +20,12 @@ def user_add(req):
 
     new_user.password = generate_password_hash(new_user.password).decode('utf8')
 
-    try:
-        db.session.add(new_user)
-        db.session.commit()
-        return jsonify({"message": "user created", "result": app_user_schema.dump(new_user)}), 201
-    except Exception as e:
-        print(e)
-        db.session.rollback()
-        return jsonify({"message": "could not create user"}), 400
+    db.session.add(new_user)
+    db.session.commit()
+    return jsonify({"message": "user created", "result": app_user_schema.dump(new_user)}), 201
     
     
-
+@auth
 def users_get_all():
     users_query = db.session.query(AppUsers).all()
 
@@ -40,21 +35,23 @@ def users_get_all():
     return jsonify({"message": "users found", "results": app_users_schema.dump(users_query)}), 200
 
 
-@auth
-def user_get_by_id(user_id):
+@authenticate_return_auth
+def user_get_by_id(req, user_id, auth_info):
     if not validate_uuid4(user_id):
         return jsonify({"message": "cannot get user without a valid uuid"}), 400
 
     user_query = db.session.query(AppUsers).filter(AppUsers.user_id == user_id).first()
 
-    if not user_query:
-        return jsonify({"message": "user does not exist"}), 404
-    
-    return jsonify({"message": "user found", "result": app_user_schema.dump(user_query)}), 200
+    if user_id == str(auth_info.user.user_id) or auth_info.user.role == 'super-admin':
+        if not user_query:
+            return jsonify({"message": "user does not exist"}), 404
+        return jsonify({"message": "user found", "result": app_user_schema.dump(user_query)}), 200
+    else:
+        return jsonify({"message": "unauthorized"}), 403
 
 
-@auth
-def user_update_by_id(req, user_id):
+@authenticate_return_auth
+def user_update_by_id(req, user_id, auth_info):
     post_data = req.form if req.form else req.json
 
     if not validate_uuid4(user_id):
@@ -64,34 +61,34 @@ def user_update_by_id(req, user_id):
 
     populate_object(user_query, post_data)
 
-    if not user_query:
-        return jsonify({"message": "user does not exist"}), 404
-    
-    try:
+    if user_id == str(auth_info.user.user_id) or auth_info.user.role == 'super-admin':
+        if not user_query:
+            return jsonify({"message": "user does not exist"}), 404
         db.session.commit()
         return jsonify({"message": "user updated", "result": app_user_schema.dump(user_query)}), 201
-    except Exception as e:
-        print(e)
-        db.session.rollback()
-        return jsonify({"message": "could not update user"}), 400
+    else:
+        return jsonify({"message": "unauthorized"}), 403
 
 
-@auth
-def user_delete_by_id(user_id):
+@authenticate_return_auth
+def user_delete_by_id(req, user_id, auth_info):
     if not validate_uuid4(user_id):
         return jsonify({"message": "cannot update user without a valid uuid"}), 400
 
     user_query = db.session.query(AppUsers).filter(AppUsers.user_id == user_id).first()
 
-    if not user_query:
-        return jsonify({"message": "user does not exist"}), 404
-    
-    try:
-        db.session.delete(user_query)
-        db.session.commit()
-        return jsonify({"message": "the user was deleted", "result": app_user_schema.dump(user_query)}), 201
-    except Exception as e:
-        print(e)
-        db.session.rollback()
-        return jsonify({"message": "could not delete user"}), 400
+    if user_id == str(auth_info.user.user_id) or auth_info.user.role == 'super-admin':
+        if not user_query:
+            return jsonify({"message": "user does not exist"}), 404
+        
+        try:
+            db.session.delete(user_query)
+            db.session.commit()
+            return jsonify({"message": "the user was deleted", "result": app_user_schema.dump(user_query)}), 201
+        except Exception as e:
+            print(e)
+            db.session.rollback()
+            return jsonify({"message": "could not delete user"}), 400
+    else:
+        return jsonify({"message": "unauthorized"}), 403
